have notification list as its own component
